Update installation.rst
The automatic process does not work no kali linux. The manual process is full of errors that need to be fixed as they prevent the source from being packaged. There are several lines containing 'rm -f' instead of 'rm -rf' within the Makefile as well as in debian/rules. The directory which extracts the source initially has to be removed, otherwise dpkg-source complains about the directory it's about to extract the source exists.   The dpkg-buildpackage command requires additional parameters to sign the package (-uc -us)