Fix reading burplog payloads
The burplog.py payload type improperly passes an str as the response_content to FuzzRequest::update_from_raw_http
as the method is expecting a bytes object. However, the method is not consistent as it expects str for the
request content (raw parameter) and a  bytes object for the response_content parameter. The method now won't try
to decode str objects.